dom/tests/mochitest/pointerlock/file_MouseEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602 - Mouse Events Allowed and Suppressed.</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #fs > div { width:100px; height:100px; }
    #inner { width:50px; height:100%; }
  </style>
</head>
<body onload="setup();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>
  <p id="display"></p>
<div id="content">
  <div id="fs">
    <div id="outer" style="background-color:Gray;">
      <div id="inner" style="background-color:Red;margin:auto;"></div>
    </div>
    <p id="outerResult">000000000</p>
    <p id="innerResult">000000000</p>
  </div>
  <pre id="test">
    <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Test will check to make sure that the following mouse events can be
       * executed prior to mouse lock and after mouse lock.
       *  - mouseover, mouseout, mouseenter, mouseleave
       *  - click, mousedown, mouseup, DOMMouseScroll
       * Mousedown and mouseup tests left, middle, and right mouse buttons.
       */
      SimpleTest.waitForExplicitFinish();
      const RIGHT_MOUSE_CLICK = 2;
      const MIDDLE_MOUSE_CLICK = 1;
      const LEFT_MOUSE_CLICK = 0;
      var moved = 0;
      var outer = document.getElementById('outer');
      var inner = document.getElementById('inner');
      var outerResult = document.getElementById('outerResult');
      var innerResult = document.getElementById('innerResult');
      var pointer = navigator.mozPointer;

      document.addEventListener('mozfullscreenchange', mouseClicks, false);

      function mouseClicks() {
        if (document.mozFullScreen && document.mozFullScreenElement === outer) {          
          outer.addEventListener('mousemove', mouseTests, false);
          pointer.lock(outer, mouseActions);
        } else {
          mouseActions();
        }
      }

      function mouseActions() {
        outerResult.innerHTML = "000000000";
        innerResult.innerHTML = "000000000";
        sendMouseEvent({ type: 'click' }, inner, window);
        synthesizeMouseAtCenter(outer, { type: 'mousedown',
          button: LEFT_MOUSE_CLICK }, window);
        synthesizeMouseAtCenter(outer, { type: 'mouseup',
          button: LEFT_MOUSE_CLICK }, window);
        synthesizeMouseAtCenter(outer, { type: 'mousedown',
          button: RIGHT_MOUSE_CLICK }, window);
        synthesizeMouseAtCenter(outer, { type: 'mouseup',
          button: RIGHT_MOUSE_CLICK }, window);
        synthesizeMouseAtCenter(outer, { type: 'mousedown',
          button: MIDDLE_MOUSE_CLICK }, window);
        synthesizeMouseAtCenter(outer, { type: 'mouseup',
          button: MIDDLE_MOUSE_CLICK }, window);
        var rect = outer.getBoundingClientRect();
        synthesizeMouseScroll(outer, rect.width/2, rect.height/2,
          {axis:"vertical", delta:10, type:"DOMMouseScroll"} , window);
        synthesizeMouse(outer, rect.width/2 + 1, rect.height/2 + 1,
          { type: 'mousemove' }, window);
      }

      function mouseTests() {
        if (pointer.isLocked) {
          todo_isnot(innerResult.innerHTML.substr(0,1), '1',
            "Right mouseup event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(1,1), '1',
            "Middle mouseup event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(2,1), '1',
            "Left mouseup event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(3,1), '1',
            "Right mousedown event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(4,1), '1',
            "Middle mousedown event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(5,1), '1',
            "Left mousedown event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(6,1), '1',
            "DOMMouseScroll event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(7,1), '1',
            "The click event was sent to an unlocked element.");
          todo_isnot(innerResult.innerHTML.substr(8,1), '1',
            "mousemove event was sent to an unlocked element.");
          is(outerResult.innerHTML.substr(0,1), '1',
            "The locked element did not get a Right mouseup event.");
          is(outerResult.innerHTML.substr(1,1), '1',
            "The locked element did not get a Middle mouseup event.");
          is(outerResult.innerHTML.substr(2,1), '1',
            "The locked element did not get a Left mouseup event.");
          is(outerResult.innerHTML.substr(3,1), '1',
            "The locked element did not get a Right mousedown event.");
          is(outerResult.innerHTML.substr(4,1), '1',
            "The locked element did not get a Middle mousedown event.");
          is(outerResult.innerHTML.substr(5,1), '1',
            "The locked element did not get the mousedown event.");
          is(outerResult.innerHTML.substr(6,1), '1',
            "The mouse locked element did not get a DOMMouseScroll event.");
          is(outerResult.innerHTML.substr(7,1), '1',
            "The mouse locked element did not get a click event.");
          is(outerResult.innerHTML.substr(8,1), '1',
            "The locked element did not get a mousemove event.");
          pointer.unlock();
          document.mozCancelFullScreen();
        } else {
          is(innerResult.innerHTML.substr(0,1), '1',
            "Right mouseup event did not fire normally.");
          is(innerResult.innerHTML.substr(1,1), '1',
            "Middle mouseup event did not fire normally.");
          is(innerResult.innerHTML.substr(2,1), '1',
            "Left mouseup event did not fire normally.");
          is(innerResult.innerHTML.substr(3,1), '1',
            "Right mousedown event did not fire normally.");
          is(innerResult.innerHTML.substr(4,1), '1',
            "Middle mousedown event did not fire normally.");
          is(innerResult.innerHTML.substr(5,1), '1',
            "Left mousedown event did not fire normally.");
          is(innerResult.innerHTML.substr(6,1), '1',
            "DOMMouseScroll event did not fire normally.");
          is(innerResult.innerHTML.substr(7,1), '1',
            "Click event did not fire normally.");
          is(innerResult.innerHTML.substr(8,1), '1',
            "mousemove event did not fire normally.");
          is(outerResult.innerHTML.substr(0,1), '1',
            "Right mouseup event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(1,1), '1',
            "Middle mouseup event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(2,1), '1',
            "Left mouseup event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(3,1), '1',
            "Right mousedown event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(4,1), '1',
            "Middle mousedown event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(5,1), '1',
            "Left mousedown event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(6,1), '1',
            "DOMMouseScroll event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(7,1), '1',
            "Click event did not bubble to the outer element.");
          is(outerResult.innerHTML.substr(8,1), '1',
            "mousemove event did not bubble to the outer element.");
          outer.removeEventListener('mousemove', mouseTests, false);
          SimpleTest.finish();
        }
      }

      function setup() {

        // Events that should be there regardless of mouse state
        outer.addEventListener('click', function() {
          outerResult.innerHTML = outerResult.innerHTML.substr(0,7) + '1' +
          outerResult.innerHTML.substr(8);
        },false);

        outer.addEventListener('mousemove', function() {
          outerResult.innerHTML = outerResult.innerHTML.substr(0,8) + '1';
        },false);

        outer.addEventListener('DOMMouseScroll', function() {
          outerResult.innerHTML = outerResult.innerHTML.substr(0,6) +
          '1' + outerResult.innerHTML.substr(7);
        },false);

        outer.addEventListener('mousedown', function(e) {
          if (e) {
            switch (e.button) {
              case LEFT_MOUSE_CLICK:
                outerResult.innerHTML = outerResult.innerHTML.substr(0,5) +
                '1' + outerResult.innerHTML.substr(6);
                break;
              case MIDDLE_MOUSE_CLICK:
                outerResult.innerHTML = outerResult.innerHTML.substr(0,4) +
                '1' + outerResult.innerHTML.substr(5);
                break;
              case RIGHT_MOUSE_CLICK:
                outerResult.innerHTML = outerResult.innerHTML.substr(0,3) +
                '1' + outerResult.innerHTML.substr(4);
                break;
            }
          }
        },false);

        outer.addEventListener('mouseup', function(e) {
          if (e) {
            switch (e.button) {
              case LEFT_MOUSE_CLICK:
                outerResult.innerHTML = outerResult.innerHTML.substr(0,2) +
                '1' + outerResult.innerHTML.substr(3);
                break;
              case MIDDLE_MOUSE_CLICK:
                outerResult.innerHTML = outerResult.innerHTML.substr(0,1) +
                '1' + outerResult.innerHTML.substr(2);
                break;
              case RIGHT_MOUSE_CLICK:
                outerResult.innerHTML = '1' + outerResult.innerHTML.substr(1);
                break;
            }
          }
        },false);

        inner.addEventListener('click', function() {
          innerResult.innerHTML = innerResult.innerHTML.substr(0,7) + '1' +
          innerResult.innerHTML.substr(8);
        },false);

        inner.addEventListener('DOMMouseScroll', function() {
          innerResult.innerHTML = innerResult.innerHTML.substr(0,6) +
          '1' + innerResult.innerHTML.substr(7);
        },false);

        inner.addEventListener('mousedown', function(e) {
          if (e) {
            switch (e.button) {
              case LEFT_MOUSE_CLICK:
                innerResult.innerHTML = innerResult.innerHTML.substr(0,5) +
                '1' + innerResult.innerHTML.substr(6);
                break;
              case MIDDLE_MOUSE_CLICK:
                innerResult.innerHTML = innerResult.innerHTML.substr(0,4) +
                '1' + innerResult.innerHTML.substr(5);
                break;
              case RIGHT_MOUSE_CLICK:
                innerResult.innerHTML = innerResult.innerHTML.substr(0,3) +
                '1' + innerResult.innerHTML.substr(4);
                break;
            }
          }
        },false);

        inner.addEventListener('mouseup', function(e) {
          if (e) {
            switch (e.button) {
              case LEFT_MOUSE_CLICK:
                innerResult.innerHTML = innerResult.innerHTML.substr(0,2) +
                '1' + innerResult.innerHTML.substr(3);
                break;
              case MIDDLE_MOUSE_CLICK:
                innerResult.innerHTML = innerResult.innerHTML.substr(0,1) +
                '1' + innerResult.innerHTML.substr(2);
                break;
              case RIGHT_MOUSE_CLICK:
                innerResult.innerHTML = '1' + innerResult.innerHTML.substr(1);
                break;
            }
          }
        },false);

        inner.addEventListener('mousemove', function() {
          innerResult.innerHTML = innerResult.innerHTML.substr(0,8) + '1';
        },false);

        startTests();
      }

      function startTests() {
        SimpleTest.waitForFocus(function() {
          outer.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
